+++ src/Transformers/ZScaleStandardizer.php
@@ -10,8 +10,8 @@ use RuntimeException;
 /**
  * Z Scale Standardizer
  *
- * A way of centering and scaling a sample matrix by computing the Z Score for
- * each feature.
+ * A method of centering and scaling a dataset such that it has 0 mean and unit
+ * variance (Z Score).
  * 
  * References:
  * [1] T. F. Chan et al. (1979). Updating Formulae and a Pairwise Algorithm for